4-5 tracking testing, pre-gui
Here is the code i've been using to test out different values for sX and sY. 
I've been monitoring the changes in the mean of u and v based on the changes to the search range. 
I was looking at the e_r_lat2-2 file on the frames 66-78 since i can see a clear down-left and up-left movement. For whatever reason the mean u, which i assume should negative from visual observation, ends up being positive in most cases.